backend/requirements.txt
fastapi
uvicorn
python-multipart
spacy
pdfminer.six
python-docx
psycopg2-binary
pydantic